+++ example/test_continue.py
@@ -1,5 +1,0 @@
-for _ in range(10):
-    # continue
-    pass
-    print(1)
-