+++ thrift4DL/server/TModelPoolServer.py
@@ -1,3 +1,8 @@
+# Copyright (c) 2019 congvm
+#
+# This software is released under the MIT License.
+# https://opensource.org/licenses/MIT
+
 from multiprocessing import Manager
 import time
 import multiprocessing
@@ -8,12 +13,11 @@ from thrift.protocol import TBinaryProtocol
 import warnings
 import numpy as np
 from thrift.Thrift import TType, TMessageType, TApplicationException
-
-multiprocessing.allow_connection_pickling()
-
+import random
 IDLE_QUEUE_BLOCK_TIME_SEC = 10
 
 __all__ = ['TModelPoolServer']
+
 
 class TModelPoolServerBase():
     def __init__(self):
@@ -22,8 +26,12 @@ class TModelPoolServerBase():
     def serve(self):
         pass
 
-class TModelPoolServerV1():
-    def __init__(self, host, port, handler_cls, model_path, gpu_ids, mem_fractions, batch_infer_size=1, batch_group_timeout=10, verbose=True, logger=None):
+
+class TModelPoolServerV1(TModelPoolServerBase):
+    def __init__(self, host, port, handler_cls,
+                 model_path, gpu_ids, mem_fractions,
+                 batch_infer_size=1, batch_group_timeout=1,
+                 verbose=True, logger=None):
         self.handler_cls = handler_cls
         self.model_path = model_path
         self.gpu_ids = gpu_ids
@@ -31,12 +39,14 @@ class TModelPoolServerV1():
         self.host = host
         self.port = port
         self.socket = TSocket.TServerSocket(host=self.host, port=self.port)
-        self.client_queue = multiprocessing.JoinableQueue()
         self.batch_infer_size = batch_infer_size
         self.batch_group_timeout = batch_group_timeout
         self.handlers = []
         self.is_running = False
         self.verbose = verbose
+        self.client_queue = multiprocessing.Queue()
+        if self.verbose:
+            self.print_server_info()
 
     def print_server_info(self):
         import pprint
@@ -74,12 +84,50 @@ class TModelPoolServerV1():
             except (SystemExit, KeyboardInterrupt):
                 break
             except Exception as err:
-                tb = traceback.format_exc()
-                print(tb)
+                print(traceback.format_exc())
+
+
+class TModelPoolServerV2(TModelPoolServerV1):
+    def prepare(self):
+        self.list_clients = []
+
+        for i in range(len(self.gpu_ids)):
+            client_queue = multiprocessing.Queue()
+            wrk = self.handler_cls(model_path=self.model_path,
+                                   gpu_id=self.gpu_ids[i],
+                                   mem_fraction=self.mem_fractions[i],
+                                   client_queue=client_queue,
+                                   batch_infer_size=self.batch_infer_size,
+                                   batch_group_timeout=self.batch_group_timeout)
+            wrk.daemon = True
+            wrk.start()
+            self.handlers.append(wrk)
+            self.list_clients.append(client_queue)
+
+        self.is_running = True
+        if self.verbose:
+            self.print_server_info()
+
+    def serve(self):
+        self.prepare()
+        self.socket.listen()
+        print("Service started")
+        while self.is_running:
+            try:
+                client = self.socket.accept()
+                if not client:
+                    continue
+                rand_idx = random.randint(0, len(self.list_clients) - 1)
+                self.list_clients[rand_idx].put(client)
+            except (SystemExit, KeyboardInterrupt):
+                break
+            except Exception as err:
+                print(traceback.format_exc())
+
 
 class TModelPoolServer():
     def __init__(self, *args, **kwargs):
-        self.__server = TModelPoolServerV1(*args, **kwargs)
+        self.__server = TModelPoolServerV2(*args, **kwargs)
 
     def __getattr__(self, name):
         return getattr(self.__server, name)