+++ thrift4DL/server/base_handler.py
@@ -1,5 +1,9 @@
+# Copyright (c) 2019 congvm
+# 
+# This software is released under the MIT License.
+# https://opensource.org/licenses/MIT
 
-from .Thrift4DLService import Receiver, Deliver
+from .connectors import Receiver, Deliver
 import multiprocessing
 from .ttypes import TVisionResult
 from thrift.Thrift import TType, TMessageType, TApplicationException