+++ thrift4DL/server/connectors.py
@@ -1,3 +1,8 @@
+# Copyright (c) 2019 congvm
+# 
+# This software is released under the MIT License.
+# https://opensource.org/licenses/MIT
+
 
 import sys
 import logging
@@ -25,6 +30,7 @@ class Receiver():
 
         self._processMap = {}
         self._processMap["predict"] = Receiver.process_predict
+        self._processMap["ping"] = Receiver.process_ping
 
     def get_connection(self, client):
         itrans = self._iptranfac.getTransport(client)
@@ -65,28 +71,56 @@ class Receiver():
             otrans.close()
             return None
         else:
-            return self.process_predict(connection_info)
+            return self._processMap[name](self, connection_info) 
 
-    def parse_args(self, iprot):
+    def parse_predict_args(self, iprot):
         args = predict_args()
         args.read(iprot)
         iprot.readMessageEnd()
         return args.image_binary
 
     def process(self, client):
+        connection_info = None
         try:
             connection_info = self.get_connection(client)
             connection_info = self.validate(connection_info)
-            return connection_info
         except Exception as e:
             print(traceback.format_exc())
+        return connection_info
 
     def process_predict(self, connection_info):
-        image_binary = self.parse_args(connection_info['iprot'])
+        image_binary = self.parse_predict_args(connection_info['iprot'])
         result = predict_result()
         connection_info['image_binary'] = image_binary
         connection_info['result'] = result
         return connection_info
+
+    def process_ping(self, connection_info):
+        seqid = connection_info['seqid']
+        iprot = connection_info['iprot']
+        oprot = connection_info['oprot']
+        args = ping_args()
+        args.read(iprot)
+        iprot.readMessageEnd()
+        result = ping_result()
+        try:
+            print("ping")
+            msg_type = TMessageType.REPLY
+        except TTransport.TTransportException:
+            raise
+        except TApplicationException as ex:
+            logging.exception('TApplication exception in handler')
+            msg_type = TMessageType.EXCEPTION
+            result = ex
+        except Exception:
+            logging.exception('Unexpected exception in handler')
+            msg_type = TMessageType.EXCEPTION
+            result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
+        oprot.writeMessageBegin("ping", msg_type, seqid)
+        result.write(oprot)
+        oprot.writeMessageEnd()
+        oprot.trans.flush()
+        return None
 
 
 class Deliver():